requirements.txt
pandas
pyautogui
selenium
numpy
opencv-python
Pillow
